gymgenius/backend/requirements.txt
# GymGenius Backend Requirements
# Python 3.11+

# ============================================================================
# Core Web Framework
# ============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# ============================================================================
# AI Provider SDKs
# ============================================================================
openai==1.57.0
google-generativeai==0.8.3

# ============================================================================
# Security & Rate Limiting
# ============================================================================
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ============================================================================
# Database
# ============================================================================
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# ============================================================================
# Testing
# ============================================================================
pytest==8.3.5
pytest-asyncio==0.25.2
pytest-cov==6.1.1
httpx==0.27.0

# ============================================================================
# Code Quality & Linting
# ============================================================================
black==24.1.1
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0
bandit==1.7.6

# ============================================================================
# Utilities
# ============================================================================
python-dotenv==1.0.1
